refactor(app): migrate delete-objects and listing flows
